requirements.txt
altair==5.5.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
decorator==5.1.1
duckdb==1.1.3
exceptiongroup==1.2.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
google-auth-oauthlib==1.2.1
gspread==5.12.4
gspread-dataframe==4.0.0
gspread-formatting==1.2.0
gspread-pandas==3.3.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.18.3
numpy==2.2.0
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
protobuf==5.29.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
Pygments==2.18.0
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
six==1.17.0
smmap==5.0.1
sql_metadata==2.15.0
sqlparse==0.5.3
st-gsheets-connection==0.1.0
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
watchdog==6.0.0
